g4-common: "fix" bmhb scheduler message
hexedited to hide the following:

> E ThermalEngine: [BMHB_GOV] read failed VS_TEMP

The reason for that message is that thermal-engine is
trying to access:

/sys/module/lge_monitor_thermal/parameters/vs_temp

which does not exists.
BMHD seems to be a ref to the fair scheduler as the same
code is trying to access:

/proc/sys/kernel/sched_bmhb_cs

which is referenced in fair.c (kernel). IMHO the whole
thing here is useless so com'on, lets hide it.. :P

New log message:

> E ThermalEngine: vs_temp skipped (that's ok)

Signed-off-by: steadfasterX <steadfasterX@gmail.com>